refactor(frontend): update permissions dialogs
